Created an app to solve Wordle
Created using Kivy for Python. Takes in five letter words from valid-wordle-words.txt in this repo. Displays a simple GUI on launch for a user to fill in their guesses up to this point for the daily Wordle challenge. The app then filters out invalid guesses based on the user's guesses to return any valid words. This is done by pressing a submit button to call the functions which filter letters based on whether they are green, yellow, or grey. Testing on this was conducted on July 31st using that day's Wordle guesses from a family member. The app displayed "cramp", the correct word for the day, based on the guesses already input. Make sure your machine runs Python and Kivy before attempting to run this on your end.